Code review comments addressed
